requirements.txt
# Core AI/ML
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# API and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0

# Memory and Storage (for legacy system)
chromadb>=0.4.0
sentence-transformers>=2.2.0  # Only for old system, not used in pure Indi

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0.0
